Figures all we needed to do was follow Next's example.
