+++ .gitignore
@@ -1,19 +1,19 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
 # dependencies
-node_modules
-.pnp/
+/node_modules
+/.pnp
 .pnp.js
 
 # testing
 /coverage
 
 # next.js
-.next/
-out/
+/renderer/.next/
+/renderer/out/
 
 # production
-/build
+/main
 /dist
 
 # misc
