+++ electron/electron-next.d.ts
@@ -1,0 +1,11 @@
+declare module 'electron-next' {
+  interface Directories {
+    production: string
+    development: string
+  }
+
+  export default function (
+    directories: Directories | string,
+    port?: number
+  ): Promise<void>
+}
