+++ renderer/components/box.tsx
