+++ renderer/components/window.tsx
