+++ renderer/public/favicon.ico
