+++ renderer/public/outro-hero.png
